Made disable_xdebug.sh script smarter to only disable xdebug when needed

+++ bin/.travis/disable_xdebug.sh
@@ -1,4 +1,6 @@
 #! /bin/bash
 
-echo "> Disable xdebug";
-phpenv config-rm xdebug.ini ;
+if [[ "$SYMFONY_DEBUG" == "" && "$TRAVIS_PHP_VERSION" != "" && "$TRAVIS_PHP_VERSION" != "hhvm" ]]; then
+    echo "> Disable xdebug";
+    phpenv config-rm xdebug.ini ;
+fi
